Data models in shared module
