--- Frontend/iosApp/Develop/DevelopApp.swift
@@ -1,18 +1,0 @@
-//
-//  DevelopApp.swift
-//  Develop
-//
-//  Created by Michał Lisicki on 27/04/2025.
-//  Copyright © 2025 orgName. All rights reserved.
-//
-
-import SwiftUI
-
-@main
-struct DevelopApp: App {
-    var body: some Scene {
-        WindowGroup {
-            ContentView()
-        }
-    }
-}
