--- Frontend/iosApp/iosApp/View/Points.swift
@@ -1,11 +1,0 @@
-//
-//  Points.swift
-//  iosApp
-//
-//  Created by Michał Lisicki on 27/04/2025.
-//  Copyright © 2025 orgName. All rights reserved.
-//
-
-import SwiftUI
-
-
