+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/ActivityDto.kt
@@ -1,0 +1,8 @@
+package com.polan.baklava.model
+
+data class ActivityDto(
+        var id: Long,
+        var student: String,
+        var source: ExerciseDto,
+        var activityValue: Double
+)
