+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/ClassGroupDto.kt
@@ -1,0 +1,10 @@
+package com.polan.baklava.model
+
+data class ClassGroupDto(
+    var id: Long,
+    var name: String,
+    var instructor: String,
+    var creator: String,
+    var classTimes: Set<ClassTime>,
+    var sessions: Set<ClassSessionDto>
+)
