+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/ClassSessionDto.kt
@@ -1,0 +1,8 @@
+package com.polan.baklava.model
+import kotlinx.datetime.LocalDateTime
+
+data class ClassSessionDto(
+    var id: Long,
+    var classDate: LocalDateTime,
+    var classGroupName: String
+)
