+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/ClassTime.kt
@@ -1,0 +1,9 @@
+package com.polan.baklava.model
+
+import kotlinx.datetime.DayOfWeek
+import kotlinx.datetime.LocalTime
+
+class ClassTime (
+    val dayOfWeek: DayOfWeek,
+    val time: LocalTime
+)
