+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/DeclarationDto.kt
@@ -1,0 +1,10 @@
+package com.polan.baklava.model
+import kotlinx.datetime.LocalDateTime
+
+data class DeclarationDto(
+        val id: Long,
+        val declarationDate: LocalDateTime,
+        val declarationStatus: DeclarationStatus,
+        val exercise: ExerciseDto,
+        val student: String
+)
