+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/DeclarationStatus.kt
@@ -1,0 +1,8 @@
+package com.polan.baklava.model
+
+enum class DeclarationStatus {
+    WAITING,
+    CANCELLED,
+    REJECTED,
+    APPROVED
+}
