+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/NewGroupDto.kt
@@ -1,0 +1,9 @@
+package com.polan.baklava.model
+
+data class NewGroupDto(
+    var id: Long,
+    var name: String,
+    var instructor: String,
+    var creator: String,
+    var classTimes: Set<ClassTime>
+)
