+++ Frontend/shared/src/commonMain/kotlin/com/polan/baklava/model/UserDto.kt
@@ -1,0 +1,7 @@
+package com.polan.baklava.model
+
+data class UserDto(
+        val email: String,
+        val name: String,
+        val surname: String
+)
